Put torch stuff together with torch
